requirements.txt
streamlit
Flask
twilio
openai
azure-cognitiveservices-speech
